atualiza README, implementa teste
