+++ src/assignCities.js
@@ -1,0 +1,7 @@
+const data = require('../data/countries_data')
+
+const assignCities = () => {
+  // desenvolva sua função aqui
+}
+
+module.exports = assignCities;
