--- src/assingCities.js
@@ -1,12 +1,0 @@
-const data = require('../data/countries_data')
-
-/* Tudo bem, Dev?
-Nosso diretor disse que agora teremos novas classificações em relações as cidades dos países
-São três classificações:
-Megalopolis: Cidades que tenham habitantes iguais ou superiores a 15 milhões
-Megacity: Cidades que tenham habitantes iguais ou superiores a 10 milhões e inferiores a 15 milhões
-Metropolis: Cidades que tenham habitantes inferiores a 10 milhões
-Você deverá elaborar uma função que, ao receber um país como parâmetro, retorne um objeto contendo
-as cidades daquele país classificada da seguinte forma:
-ex.: { megalopolis: 3, megacity: 6, metropolis: 4 } */
-
