+++ test/getFirstCountry.test.js
@@ -1,0 +1,2 @@
+const data = require('../data/countries_data')
+
