+++ test/isOnTheList.test.js
@@ -1,0 +1,1 @@
+const data = require('../data/countries_data')
